fixing merge conflicts and testing out modal/follow
