+++ react-app/src/components/ViewPostModal/index.js
@@ -1,21 +1,21 @@
-import React, { useState } from "react";
-import { Modal } from "../../context/Modal";
-import SinglePost from "../SinglePost";
-import SinglePostModal from "./SinglePostModal";
+// import React, { useState } from "react";
+// import { Modal } from "../../context/Modal";
+// import SinglePost from "../SinglePost";
+// import SinglePostModal from "./SinglePostModal";
 
-function ViewPostModal({ post, isOpen, onClose }) {
-  // const [showModal, setShowModal] = useState(false);
+// function ViewPostModal({ post, isOpen, onClose }) {
+//   // const [showModal, setShowModal] = useState(false);
 
-  return (
-    <>
-      {/* <button onClick={() => setShowModal(true)}></button> */}
-      {isOpen && (
-        <Modal onClose={onClose}>
-          <SinglePost post={post} />
-        </Modal>
-      )}
-    </>
-  );
-}
+//   return (
+//     <>
+//       {/* <button onClick={() => setShowModal(true)}></button> */}
+//       {isOpen && (
+//         <Modal onClose={onClose}>
+//           <SinglePost post={post} />
+//         </Modal>
+//       )}
+//     </>
+//   );
+// }
 
-export default ViewPostModal;
+// export default ViewPostModal;
