+++ react-app/src/context/Modal.css
@@ -7,6 +7,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
+    z-index: 5;
   }
 
   #modal-background {
@@ -25,4 +26,5 @@
     background-color:white;
     border-radius: 16px;
     padding: 15px;
+    z-index: 6;
   }
